refactor(builtin): the glob and grep tests move beside the write tests
crates/nika-builtin/src/file.rs crossed the 1500-line wall (1543) once
the left-out-directory report and #1504's write teaching met on the same
file. The glob and grep tests (376 lines) move to
file/glob_grep_tests.rs, the sibling pattern of file/write_tests.rs, as
a pure move; the value-only `glob` door they read travels with them.
398 lib tests unchanged; file.rs 1164 lines.
